+++ database/model.go
@@ -16,16 +16,16 @@ type Database struct {
 type ImageRow struct {
 	Signature     string  `field:"signature" json:"signature"`
 	Extension     string  `field:"extension" json:"extension"`
-	ImageId       uint    `field:"image_id" json:"image_id"`
-	Favorites     uint    `field:"favorites" json:"favorites"`
+	ImageId       int64   `field:"image_id" json:"image_id"`
+	Favorites     int64   `field:"favorites" json:"favorites"`
 	DominantColor string  `field:"dominant_color" json:"dominant_color"`
 	Source        *string `field:"source" json:"source"`
 	UploadedAt    string  `field:"uploaded_at" json:"uploaded_at"`
 	LikedAt       *string `field:"liked_at" json:"liked_at"`
 	IsNsfw        bool    `field:"is_nsfw" json:"is_nsfw"`
-	Width         uint    `field:"width" json:"width"`
-	Height        uint    `field:"height" json:"height"`
-	TagId         uint    `field:"tag_id" json:"tag_id"`
+	Width         int64   `field:"width" json:"width"`
+	Height        int64   `field:"height" json:"height"`
+	TagId         int64   `field:"tag_id" json:"tag_id"`
 	Name          string  `field:"name" json:"name"`
 	Description   string  `field:"description" json:"description"`
 	TagIsNsfw     bool    `field:"tag_is_nsfw" json:"tag_is_nsfw"`
@@ -37,7 +37,7 @@ type ImageRows struct {
 	Rows          []ImageRow
 }
 
-func (ir ImageRows) GetImage(ImageId uint) models.Image {
+func (ir ImageRows) GetImage(ImageId int64) models.Image {
 	var image models.Image
 	for _, im := range ir.Rows {
 		if im.ImageId == ImageId {
@@ -67,15 +67,15 @@ func (ir ImageRows) GetImage(ImageId uint) models.Image {
 
 // PermissionsInformation a struct used to retrieve and a row from the database
 type PermissionsInformation struct {
-	UserId   uint
-	TargetId *uint
+	UserId   int64
+	TargetId *int64
 	Position int
 	Name     string
 }
 
 type ReportRes struct {
-	ImageId     uint    `json:"image_id" field:"image_id" `
-	AuthorId    uint    `json:"author_id" field:"author_id"`
+	ImageId     int64   `json:"image_id" field:"image_id" `
+	AuthorId    int64   `json:"author_id" field:"author_id"`
 	Description *string `json:"description" field:"description"`
 	Existed     bool    `json:"existed" field:"existed"`
 }
